Swap tl4j with guava-testlib
guava-testlib has all the tl4j stuff, except for TearDownTestCase (aka
TDTC).

The only production code dependency to TDTC was from the helper class
GuiceBerryJunit3TestCase. To eliminate this dependency, the code from
TDTC that performs the teardown (by using the TearDownStack) was
inlined in that class, and GuiceBerryJunit3TestCase, instead of
extending TDTC, now implements TearDownAccepter. Everything should
still work exactly as before _except_ if someone either casts
(compile-time breakage) or typecasts (runtime-breakage) a
GuiceBerryJunit3TestCase as a TDTC. Though doing either of these is
unlikely, this commit will bump the next release's major revision
number (i.e. to 4.0.0).

Now, the code tutorial for TDTC-based GuiceBerry still needs a TDTC
class (otherwise it will not compile). Furthermore, some GuiceBerry
tests themselves still use TDTC. For this reason, a copy of TDTC was
added to the _test_ tree, so neither of these need to change with this
CL. Since the TDTC here is in the "test" tree, it will _not_ be
released with the jar: it is a effective a private class to GuiceBerry
tests.