Fix bug in initialization order introduced by a0cea1f0f6319614fe092a339a451ad432125a83
Fix an issue introduced by
https://github.com/zorzella/guiceberry/pull/32/files that caused a
change in the initialization order -- if there was a TestWrapper used,
it would be provisioned before the GuiceBerryEnvMain.run method was
run. This is particularly problematic if the GuiceBerryEnvMain is used
to change some global static state that is then read by a singleton
binding that is injected by the test wrapper, as that singleton would
be built before the GuiceBerryEnvMain runs (i.e. before it is
initialized).